+++ aliri_actix/src/macros.rs
@@ -122,7 +122,7 @@
 /// use aliri::jwt;
 /// use aliri_actix::scope_policy;
 /// use aliri_clock::UnixTime;
-/// use aliri_oauth2::oauth2;
+/// use aliri_oauth2::{HasScope, Scope};
 /// use serde::Deserialize;
 /// use actix_web::{get, HttpResponse, Responder};
 ///
@@ -131,7 +131,7 @@
 ///     iss: jwt::Issuer,
 ///     aud: jwt::Audiences,
 ///     sub: jwt::Subject,
-///     scope: oauth2::Scope,
+///     scope: Scope,
 /// }
 ///
 /// impl jwt::CoreClaims for CustomClaims {
@@ -142,8 +142,8 @@
 ///     fn sub(&self) -> Option<&jwt::SubjectRef> { Some(&self.sub) }
 /// }
 ///
-/// impl oauth2::HasScope for CustomClaims {
-///     fn scope(&self) -> &oauth2::Scope { &self.scope }
+/// impl HasScope for CustomClaims {
+///     fn scope(&self) -> &Scope { &self.scope }
 /// }
 ///
 /// // Define our initial scope
@@ -167,7 +167,7 @@
 #[macro_export]
 macro_rules! scope_policy {
     ($i:ident/$s:ident; $($($scope:literal),*);*) => {
-      scope_policy!($i/$s(::aliri_oauth2::oauth2::BasicClaimsWithScope); $($($scope),*);*);
+      scope_policy!($i/$s(::aliri_oauth2::scope::BasicClaimsWithScope); $($($scope),*);*);
     };
     ($i:ident/$s:ident($claim:ty); $($($scope:literal),*);*) => {
         #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -199,7 +199,7 @@ macro_rules! scope_policy {
                         .or_allow(
                             ::aliri_oauth2::Scope::empty()
                             $(
-                                .and(::aliri_oauth2::oauth2::ScopeToken::from_static($scope))
+                                .and(::aliri_oauth2::scope::ScopeToken::from_static($scope))
                             )*
                         )
                     )*