+++ aliri_tower/examples/tonic.rs
@@ -3,7 +3,7 @@ use std::sync::atomic::{AtomicI32, Ordering};
 use aliri::{jwa, jwk, jwt, Jwk, Jwks, Jwt};
 use aliri_base64::Base64UrlRef;
 use aliri_clock::{Clock, DurationSecs, UnixTime};
-use aliri_oauth2::{oauth2, oauth2::HasScope, policy, scope, Authority};
+use aliri_oauth2::{policy, scope, Authority, HasScope};
 use aliri_tower::Oauth2Authorizer;
 use aliri_traits::Policy;
 use counter::{
@@ -121,7 +121,7 @@ pub struct CustomClaims {
     aud: jwt::Audiences,
     sub: jwt::Subject,
     exp: UnixTime,
-    scope: oauth2::Scope,
+    scope: scope::Scope,
 }
 
 impl jwt::CoreClaims for CustomClaims {
@@ -142,8 +142,8 @@ impl jwt::CoreClaims for CustomClaims {
     }
 }
 
-impl oauth2::HasScope for CustomClaims {
-    fn scope(&self) -> &oauth2::Scope {
+impl HasScope for CustomClaims {
+    fn scope(&self) -> &scope::Scope {
         &self.scope
     }
 }