+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This changelog is based on the format from [Keep a Changelog](https://keepachang
 
 ## [Unreleased]
 
+- (oauth2) BREAKING: Rename `oauth2` module to `scope` ([#37])
+- (oauth2) Re-export `HasScope` at crate root
+
+  [#37]: https://github.com/neoeinstein/aliri/pull/37
+
 ## [2023-03-04]
 
 - `aliri_axum` to 0.2.1